Botones de home y agregar producto agregados, pantalla agregar producto funcional, queda hacer que el producto se agregue al catalogo

+++ app/src/main/java/com/example/login/model/Carrito.kt
@@ -3,10 +3,10 @@ package com.example.login.model
 data class Carrito (
     val id : Int,
     val nombre: String,
-    val precio: Double,
+    val precio: Int,
     var cantidad: Int = 1,
     val imagen: String = ""
 ) {
-    val subtotal: Double
+    val subtotal: Int
         get() = precio * cantidad
 }
